# Conflicts:
#	src/main/java/controller/BookController.java
#	src/main/java/controller/CarController.java
#	src/main/java/controller/RootVegetableController.java
#	src/main/java/datamodels/Book.java
#	src/main/java/datamodels/Car.java
#	src/main/java/datamodels/RootVegetable.java
#	src/main/java/ioData/BookCreatorUtil.java
#	src/main/java/ioData/CarCreatorUtil.java
#	src/main/java/ioData/IOManager.java
#	src/main/java/ioData/RootVegetablesCreatorUtil.java
#	src/main/java/ioData/TestAPP.java
#	src/main/java/searchItems/BinarySearcher.java
#	src/main/java/sorters/ShellSort.java

+++ src/main/java/controller/CarController.java
@@ -37,7 +37,7 @@ public class CarController {
                     String[] parse = readerUser.create(validation, null, Controller.scanner);
                     database.add(new Car.CarBuilder()
                             .model(parse[0])
-                            .power(Double.parseDouble(parse[1]) )
+                            .power(Integer.parseInt(parse[1]))
                             .yearOfManufacture(Integer.parseInt(parse[2]))
                             .build());
                     System.out.println(reader.StringsConsole.ENTER_MORE);
